Update binomial_tree_CRR.py
Updated the _calculate_option_price method to speed up computations.

The underlying prices tree is now generated in a single step using a triangular boolean mask (np.tri) and a vectorized operation instead of a loop leading to improved performance and shorter code.

The option price tree calculation using backward induction has been simplified. This now done using a for loop that iterates backwards and update the option value for all the columns at once by leveraging array slicing. In the previous version, the nested for loop was iterating backwards though the rows and was applying the operation to each node.