Update binomial_tree_CRR.py
Updated the class to also handle American options
Changed the class name to Option_CRR.
Addition of a the opt_style parameter.
Expanded the calculation price method with an if statement depending on opt_style. If opt_style is American, the logic now updates the option tree's values with the early exercise payoff at each time step.